server: fix dynamodb ttl to be in seconds, fix build scripts

+++ server/build.ps1
@@ -1,4 +1,8 @@
-Remove-Item dist -Recurse -Force
-Remove-Item server.zip -Force
+if (Test-Path dist) {
+    Remove-Item dist -Recurse -Force
+}
+if (Test-Path server.zip) {
+    Remove-Item server.zip -Force
+}
 tsc
 7z a -tzip -r ./server.zip .\dist\*
